Use `intptr_t` etc. when casting pointers (#2811)
Assuming that, for example, `void*` can be cast to `uint32_t` and vice-versa is not portable since platforms can have different sizes. Correct types to use are `intptr_t` and `uintptr_t` which are 32-bit for all the architectures we care about, and 64-bit for modern OS builds without 32-bit compatibility mode. It also makes it clearer that values are likely cast from pointers.

+++ Sming/Components/ssl/Axtls/AxContext.cpp
@@ -89,7 +89,7 @@ Connection* AxContext::createClient(tcp_pcb* tcp)
 	auto id = session.getSessionId();
 	auto connection = new AxConnection(*this, tcp);
 	auto client =
-		ssl_client_new(context, int(connection), id ? id->getValue() : nullptr, id ? id->getLength() : 0, ssl_ext);
+		ssl_client_new(context, intptr_t(connection), id ? id->getValue() : nullptr, id ? id->getLength() : 0, ssl_ext);
 	if(client == nullptr) {
 		ssl_ext_free(ssl_ext);
 		delete connection;
@@ -105,7 +105,7 @@ Connection* AxContext::createServer(tcp_pcb* tcp)
 	assert(context != nullptr);
 
 	auto connection = new AxConnection(*this, tcp);
-	auto server = ssl_server_new(context, int(connection));
+	auto server = ssl_server_new(context, intptr_t(connection));
 	if(server == nullptr) {
 		delete connection;
 		return nullptr;
